+++ src/persistence/mod.rs
@@ -4,7 +4,7 @@ mod constants;
 mod format;
 mod load;
 mod save;
-mod state;
+mod window_state;
 
 pub use format::WindowKey;
 pub(crate) use load::get_default_state_path;
@@ -13,5 +13,5 @@ pub(crate) use load::load_all_states;
 pub(crate) use save::save_active_window_state;
 pub(crate) use save::save_all_states;
 pub(crate) use save::save_window_state;
-pub(crate) use state::SavedWindowMode;
-pub(crate) use state::WindowState;
+pub(crate) use window_state::SavedWindowMode;
+pub(crate) use window_state::WindowState;